Changed how CPU frequency is calculated.
First, use CPU topology info to associate threads to cores. Select the max
frequency for each of the cores.

Second, use core base frequency to decide whether the core is busy or idle.
If there are busy cores, calculate their average frequency and report that.
Otherwise, report average frequency of the idle cores.