Count CPU core frequencise as active if at least 100MHz above base.
Linux reports effective clocks, which means even if the core has switched
to a frequency above base for only a small fraction of the monitoring interval,
the reported frequency will be slightly above base. Sich core will be counted
as "active", despite that most of the time it operated on base clock. This
tanks the overall estimate of the frequency of the truly "active" cores.

To mitigate this, only count cores as "active" if their reported effective
clock is 100MHz above base or higher.